+++ src/component/NodeEditor/index.tsx
@@ -6,7 +6,16 @@ import NodeEditor, {
     isSwitchForm,
     getAction
 } from './NodeEditor';
+import { DEFAULT_BODY } from './constants';
 
-export { GetResultNameField, NodeEditorProps, FormProps, mapExits, isSwitchForm, getAction };
+export {
+    GetResultNameField,
+    NodeEditorProps,
+    FormProps,
+    DEFAULT_BODY,
+    mapExits,
+    isSwitchForm,
+    getAction
+};
 
 export default NodeEditor;
