+++ src/component/routers/SubflowRouter.tsx
@@ -7,24 +7,14 @@ import FlowElement from '../form/FlowElement';
 import ComponentMap from '../../services/ComponentMap';
 import { ConfigProviderContext, Type, endpointsPT } from '../../config';
 
-export interface SubflowRouterFormProps extends FormProps {
-    action: StartFlow;
-    node: Node;
-    config: Type;
-    updateRouter(node: Node, type: string, previousAction: AnyAction): void;
-    onBindWidget(ref: any): void;
-    translating: boolean;
-    getExitTranslations(): JSX.Element;
-    saveLocalizedExits(widgets: { [name: string]: any }): void;
-    ComponentMap: ComponentMap;
-}
+export type SubflowRouterProps = Partial<FormProps>;
 
-export default class SubflowRouter extends React.PureComponent<SubflowRouterFormProps> {
+export default class SubflowRouter extends React.PureComponent<SubflowRouterProps> {
     public static contextTypes = {
         endpoints: endpointsPT
     };
 
-    constructor(props: SubflowRouterFormProps, context: ConfigProviderContext) {
+    constructor(props: SubflowRouterProps, context: ConfigProviderContext) {
         super(props);
 
         this.onValid = this.onValid.bind(this);
@@ -32,73 +22,10 @@ export default class SubflowRouter extends React.PureComponent<SubflowRouterForm
 
     public onValid(widgets: { [name: string]: any }): void {
         if (this.props.translating) {
-            return this.props.saveLocalizedExits(widgets);
+            return this.props.saveLocalizations(widgets);
         }
 
-        const select = widgets.Flow as FlowElement;
-        const { name: flowName, id: flowUUID } = select.state.flow;
-
-        const newAction: StartFlow = {
-            uuid: this.props.action.uuid,
-            type: this.props.config.type,
-            flow_name: flowName,
-            flow_uuid: flowUUID
-        };
-
-        // If we're already a subflow, lean on those exits and cases
-        let exits: Exit[];
-        let cases: Case[];
-
-        const details = this.props.ComponentMap.getDetails(this.props.node.uuid);
-
-        if (details && details.type === 'subflow') {
-            ({ exits } = this.props.node);
-            ({ cases } = this.props.node.router as SwitchRouter);
-        } else {
-            // Otherwise, let's create some new ones
-            exits = [
-                {
-                    uuid: generateUUID(),
-                    name: 'Complete',
-                    destination_node_uuid: null
-                }
-            ];
-
-            cases = [
-                {
-                    uuid: generateUUID(),
-                    type: 'has_run_status',
-                    arguments: ['C'],
-                    exit_uuid: exits[0].uuid
-                }
-            ];
-        }
-
-        const newRouter: SwitchRouter = {
-            type: 'switch',
-            operand: '@child',
-            cases,
-            default_exit_uuid: null
-        };
-
-        // HACK: this should go away with modal refactor
-        let { uuid: nodeUUID } = this.props.node;
-
-        if (this.props.action.uuid === nodeUUID) {
-            nodeUUID = generateUUID();
-        }
-
-        this.props.updateRouter(
-            {
-                uuid: nodeUUID,
-                router: newRouter,
-                exits,
-                actions: [newAction],
-                wait: { type: 'flow', flow_uuid: flowUUID }
-            },
-            'subflow',
-            this.props.action
-        );
+        this.props.updateRouter();
     }
 
     public render(): JSX.Element {
@@ -113,8 +40,8 @@ export default class SubflowRouter extends React.PureComponent<SubflowRouterForm
                     ref={this.props.onBindWidget}
                     name="Flow"
                     endpoint={this.context.endpoints.flows}
-                    flow_name={this.props.action.flow_name}
-                    flow_uuid={this.props.action.flow_uuid}
+                    flow_name={(this.props.action as StartFlow).flow_name}
+                    flow_uuid={(this.props.action as StartFlow).flow_uuid}
                     required={true}
                 />
             </div>
