+++ src/component/routers/constants.ts
@@ -1,7 +1,5 @@
-export const DEFAULT_OPERAND = '@input';
 export const WAIT_LABEL = 'If the message response...';
 export const EXPRESSION_LABEL = 'If the expression...';
-export const GROUPS_OPERAND = '@contact.groups';
 export const GROUP_LABEL = "Select the group(s) you'd like to split by below";
 export const OPERATOR_LOCALIZATION_LEGEND =
     'Sometimes languages need special rules to route things properly. If a translation is not provided, the original rule will be used.';
